Fix facet count method
