+++ shared/src/db/repositories/document.rs
@@ -1,22 +1,12 @@
 use crate::{
     db::error::DatabaseError,
-    models::{AttributeFilter, DateFilter, Document, Facet, FacetValue},
+    models::{AttributeFilter, DateFilter, Document},
     SourceType,
 };
 use serde_json::Value as JsonValue;
 use sqlx::{FromRow, PgPool};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use time::{self, OffsetDateTime};
-use tracing::debug;
-
-#[derive(FromRow)]
-pub struct SearchHit {
-    #[sqlx(flatten)]
-    pub document: Document,
-    pub score: f32,
-    #[sqlx(default)]
-    pub content_snippets: Option<Vec<String>>,
-}
 
 #[derive(FromRow)]
 pub struct TitleEntry {
@@ -290,311 +280,6 @@ impl DocumentRepository {
         }
     }
 
-    pub async fn search(
-        &self,
-        query: &str,
-        source_ids: &[String],
-        content_types: Option<&[String]>,
-        attribute_filters: Option<&HashMap<String, AttributeFilter>>,
-        limit: i64,
-        offset: i64,
-        user_email: Option<&str>,
-        document_id: Option<&str>,
-        date_filter: Option<&DateFilter>,
-        person_filters: Option<&[String]>,
-        recency_boost_weight: f32,
-        recency_half_life_days: f32,
-    ) -> Result<Vec<SearchHit>, DatabaseError> {
-        if source_ids.is_empty() {
-            return Ok(vec![]);
-        }
-
-        // Filter-only search: no BM25 scoring, just apply filters sorted by recency
-        if query.trim().is_empty() {
-            return self
-                .filter_only_search(
-                    source_ids,
-                    content_types,
-                    attribute_filters,
-                    limit,
-                    offset,
-                    user_email,
-                    date_filter,
-                )
-                .await;
-        }
-
-        // We use term-centric ranking
-        // Score each query term independently across fields (best-field per term),
-        // then SUM across terms. This rewards docs matching MORE query terms.
-        // Phrase bonus is additive on top.
-
-        // Tokenize query via ParadeDB: stems, removes stopwords, ASCII-folds
-        // This matches the index tokenizer pipeline exactly.
-        let raw_terms: Vec<String> = sqlx::query_scalar(
-            "SELECT unnest($1::pdb.simple('stemmer=english', 'stopwords_language=english', 'ascii_folding=true')::text[])"
-        )
-        .bind(query)
-        .fetch_all(&self.pool)
-        .await?;
-
-        let mut seen = HashSet::new();
-        let terms: Vec<String> = raw_terms
-            .into_iter()
-            .filter(|t| seen.insert(t.clone()))
-            .take(8)
-            .collect();
-
-        // Bind params: $1 = full query, $2..$(1+N) = individual terms, then filters
-        let mut param_idx = 2 + terms.len();
-
-        let mut filters = Vec::new();
-        self.build_common_filters(
-            &mut filters,
-            &mut param_idx,
-            source_ids,
-            content_types,
-            attribute_filters,
-            user_email,
-            date_filter,
-        );
-
-        if document_id.is_some() {
-            filters.push(format!("id = ${}", param_idx));
-            param_idx += 1;
-        }
-
-        // Person filters: strict author filtering via BM25 index on metadata
-        if let Some(persons) = person_filters {
-            let conditions: Vec<String> = persons
-                .iter()
-                .map(|p| {
-                    let escaped = p.replace('\'', "''");
-                    format!("metadata ||| 'author:{escaped}'")
-                })
-                .collect();
-            if !conditions.is_empty() {
-                filters.push(format!("({})", conditions.join(" OR ")));
-            }
-        }
-
-        let common_where = if filters.is_empty() {
-            String::new()
-        } else {
-            format!(" AND {}", filters.join(" AND "))
-        };
-
-        // Per-term: best of title (default tokenizer), title (source_code tokenizer to handle
-        // CamelCase), and content
-        let mut term_branches = Vec::new();
-        for (i, _term) in terms.iter().enumerate() {
-            let term_param = format!("${}", 2 + i);
-            term_branches.push(format!(
-                "SELECT id, MAX(score) as score FROM (\
-                    SELECT id, pdb.score(id) as score FROM documents \
-                    WHERE title ||| {term_param}::pdb.boost(2){common_where} \
-                    UNION ALL \
-                    SELECT id, pdb.score(id) as score FROM documents \
-                    WHERE title::pdb.alias('title_secondary') ||| {term_param}::pdb.boost(2){common_where} \
-                    UNION ALL \
-                    SELECT id, pdb.score(id) as score FROM documents \
-                    WHERE content ||| {term_param}{common_where}\
-                ) t{i} GROUP BY id"
-            ));
-        }
-
-        // Phrase branches: best of title phrase vs content phrase (using $1 = full query)
-        let phrase_branch = format!(
-            "SELECT id, MAX(score) as score FROM (\
-                SELECT id, pdb.score(id) as score FROM documents \
-                WHERE title ### $1::pdb.slop(2)::pdb.boost(10){common_where} \
-                UNION ALL \
-                SELECT id, pdb.score(id) as score FROM documents \
-                WHERE content ### $1::pdb.slop(2)::pdb.boost(5){common_where}\
-            ) p GROUP BY id"
-        );
-
-        // When all query terms are stopwords, terms is empty — skip term_scores,
-        // rank by phrase scoring only.
-        let weight_idx = param_idx + 2;
-        let half_life_idx = param_idx + 3;
-
-        let recency_expr = format!(
-            "(1.0 + ${w}::double precision * EXP(-EXTRACT(EPOCH FROM (CURRENT_TIMESTAMP - COALESCE(\
-                CASE WHEN d.metadata->>'updated_at' IS NOT NULL \
-                     AND pg_input_is_valid(d.metadata->>'updated_at', 'timestamptz') \
-                THEN (d.metadata->>'updated_at')::timestamptz END, \
-                d.updated_at))) / (86400.0 * ${h}::double precision)))::real",
-            w = weight_idx,
-            h = half_life_idx,
-        );
-
-        let full_query = if terms.is_empty() {
-            format!(
-                r#"
-                WITH phrase_scores AS (
-                    {phrase_branch}
-                ),
-                ranked AS (
-                    SELECT ps.id, (ps.score * {recency_expr}) as score
-                    FROM phrase_scores ps
-                    JOIN documents d ON d.id = ps.id
-                    ORDER BY score DESC
-                    LIMIT ${limit_idx} OFFSET ${offset_idx}
-                )
-                SELECT r.id, r.score,
-                       d.source_id, d.external_id, d.title, d.content_id, d.content_type,
-                       d.file_size, d.file_extension, d.url,
-                       d.metadata, d.permissions, d.attributes, d.created_at, d.updated_at, d.last_indexed_at,
-                       COALESCE(snip.content_snippets, ARRAY[LEFT(d.content, 240)]) as content_snippets
-                FROM ranked r
-                JOIN documents d ON d.id = r.id
-                LEFT JOIN LATERAL (
-                    SELECT pdb.snippets(doc.content, start_tag => '**', end_tag => '**',
-                                        max_num_chars => 200, "limit" => 3, sort_by => 'score') as content_snippets
-                    FROM documents doc
-                    WHERE doc.content ||| $1 AND doc.id = r.id
-                    LIMIT 1
-                ) snip ON true
-                ORDER BY r.score DESC"#,
-                phrase_branch = phrase_branch,
-                recency_expr = recency_expr,
-                limit_idx = param_idx,
-                offset_idx = param_idx + 1,
-            )
-        } else {
-            format!(
-                r#"
-                WITH term_scores AS (
-                    {term_union}
-                ),
-                phrase_scores AS (
-                    {phrase_branch}
-                ),
-                combined AS (
-                    SELECT id, SUM(score) as token_score FROM term_scores GROUP BY id
-                ),
-                ranked AS (
-                    SELECT c.id, ((c.token_score + COALESCE(p.score, 0)) * {recency_expr}) as score
-                    FROM combined c
-                    LEFT JOIN phrase_scores p ON c.id = p.id
-                    JOIN documents d ON d.id = c.id
-                    ORDER BY score DESC
-                    LIMIT ${limit_idx} OFFSET ${offset_idx}
-                )
-                SELECT r.id, r.score,
-                       d.source_id, d.external_id, d.title, d.content_id, d.content_type,
-                       d.file_size, d.file_extension, d.url,
-                       d.metadata, d.permissions, d.attributes, d.created_at, d.updated_at, d.last_indexed_at,
-                       COALESCE(snip.content_snippets, ARRAY[LEFT(d.content, 240)]) as content_snippets
-                FROM ranked r
-                JOIN documents d ON d.id = r.id
-                LEFT JOIN LATERAL (
-                    SELECT pdb.snippets(doc.content, start_tag => '**', end_tag => '**',
-                                        max_num_chars => 200, "limit" => 3, sort_by => 'score') as content_snippets
-                    FROM documents doc
-                    WHERE doc.content ||| $1 AND doc.id = r.id
-                    LIMIT 1
-                ) snip ON true
-                ORDER BY r.score DESC"#,
-                term_union = term_branches.join("\nUNION ALL\n"),
-                phrase_branch = phrase_branch,
-                recency_expr = recency_expr,
-                limit_idx = param_idx,
-                offset_idx = param_idx + 1,
-            )
-        };
-        debug!("Full search query: {}", full_query);
-
-        let mut query_builder = sqlx::query_as::<_, SearchHit>(&full_query).bind(query);
-
-        for term in &terms {
-            query_builder = query_builder.bind(term.as_str());
-        }
-
-        query_builder = query_builder.bind(source_ids);
-
-        if let Some(ct) = content_types {
-            if !ct.is_empty() {
-                query_builder = query_builder.bind(ct);
-            }
-        }
-
-        if let Some(doc_id) = document_id {
-            query_builder = query_builder.bind(doc_id);
-        }
-
-        query_builder = query_builder
-            .bind(limit)
-            .bind(offset)
-            .bind(recency_boost_weight as f64)
-            .bind(recency_half_life_days as f64);
-
-        let results = query_builder.fetch_all(&self.pool).await?;
-
-        Ok(results)
-    }
-
-    async fn filter_only_search(
-        &self,
-        source_ids: &[String],
-        content_types: Option<&[String]>,
-        attribute_filters: Option<&HashMap<String, AttributeFilter>>,
-        limit: i64,
-        offset: i64,
-        user_email: Option<&str>,
-        date_filter: Option<&DateFilter>,
-    ) -> Result<Vec<SearchHit>, DatabaseError> {
-        let mut param_idx = 1;
-        let mut filters = Vec::new();
-        self.build_common_filters(
-            &mut filters,
-            &mut param_idx,
-            source_ids,
-            content_types,
-            attribute_filters,
-            user_email,
-            date_filter,
-        );
-
-        let where_clause = if filters.is_empty() {
-            String::new()
-        } else {
-            format!("WHERE {}", filters.join(" AND "))
-        };
-
-        let query_str = format!(
-            r#"
-            SELECT id, 0.0::real as score, source_id, external_id, title, content_id, content_type,
-                   file_size, file_extension, url,
-                   metadata, permissions, attributes, created_at, updated_at, last_indexed_at,
-                   ARRAY[LEFT(content, 240)] as content_snippets
-            FROM documents
-            {where_clause}
-            ORDER BY updated_at DESC
-            LIMIT ${limit_idx} OFFSET ${offset_idx}
-            "#,
-            where_clause = where_clause,
-            limit_idx = param_idx,
-            offset_idx = param_idx + 1,
-        );
-
-        let mut query_builder = sqlx::query_as::<_, SearchHit>(&query_str);
-
-        query_builder = query_builder.bind(source_ids);
-
-        if let Some(ct) = content_types {
-            if !ct.is_empty() {
-                query_builder = query_builder.bind(ct);
-            }
-        }
-
-        query_builder = query_builder.bind(limit).bind(offset);
-
-        let results = query_builder.fetch_all(&self.pool).await?;
-        Ok(results)
-    }
-
     pub async fn find_by_source(&self, source_id: &str) -> Result<Vec<Document>, DatabaseError> {
         let documents = sqlx::query_as::<_, Document>(
             r#"
@@ -791,103 +476,6 @@ impl DocumentRepository {
         .await?;
 
         Ok(upserted_document)
-    }
-
-    // TODO: This uses field-centric scoring (full query per field, MAX) while search() uses
-    // term-centric scoring (per-term best-field, SUM). This can produce inconsistent facet counts.
-    // Unify with search() scoring — either via a single combined query or shared scoring CTEs.
-    pub async fn get_facet_counts(
-        &self,
-        query: &str,
-        source_ids: &[String],
-        content_types: Option<&[String]>,
-        attribute_filters: Option<&HashMap<String, AttributeFilter>>,
-        user_email: Option<&str>,
-        date_filter: Option<&DateFilter>,
-    ) -> Result<Vec<Facet>, DatabaseError> {
-        if source_ids.is_empty() {
-            return Ok(vec![]);
-        }
-
-        let mut filters = Vec::new();
-        let mut param_idx = 2;
-
-        self.build_common_filters(
-            &mut filters,
-            &mut param_idx,
-            source_ids,
-            content_types,
-            attribute_filters,
-            user_email,
-            date_filter,
-        );
-
-        let common_where = if filters.is_empty() {
-            String::new()
-        } else {
-            format!(" AND {}", filters.join(" AND "))
-        };
-
-        let union_branches = [
-            format!("SELECT id, pdb.score(id) as score FROM documents WHERE title ||| $1::pdb.boost(2){common_where}"),
-            format!("SELECT id, pdb.score(id) as score FROM documents WHERE content ||| $1{common_where}"),
-            format!("SELECT id, pdb.score(id) as score FROM documents WHERE title::pdb.alias('title_secondary') ||| $1{common_where}"),
-            format!("SELECT id, pdb.score(id) as score FROM documents WHERE title ### $1::pdb.slop(2)::pdb.boost(10){common_where}"),
-            format!("SELECT id, pdb.score(id) as score FROM documents WHERE content ### $1::pdb.slop(2)::pdb.boost(5){common_where}"),
-        ];
-
-        let query_str = format!(
-            r#"
-            WITH field_scores AS (
-                {union_all}
-            ),
-            best AS (
-                SELECT id, MAX(score) as score
-                FROM field_scores
-                GROUP BY id
-            ),
-            thresholded AS (
-                SELECT id FROM best
-                WHERE score >= (SELECT MAX(score) FROM best) * 0.15
-            )
-            SELECT 'source_type' as facet, s.source_type as value, count(*) as count
-            FROM thresholded t
-            JOIN documents d ON d.id = t.id
-            JOIN sources s ON d.source_id = s.id
-            GROUP BY s.source_type
-            ORDER BY count DESC
-            "#,
-            union_all = union_branches.join("\n                UNION ALL\n                "),
-        );
-
-        let mut query_builder = sqlx::query_as::<_, (String, String, i64)>(&query_str)
-            .bind(query)
-            .bind(source_ids);
-
-        if let Some(ct) = content_types {
-            if !ct.is_empty() {
-                query_builder = query_builder.bind(ct);
-            }
-        }
-
-        let facet_rows = query_builder.fetch_all(&self.pool).await?;
-
-        let mut facets_map: std::collections::HashMap<String, Vec<FacetValue>> =
-            std::collections::HashMap::new();
-
-        for (facet_name, value, count) in facet_rows {
-            facets_map
-                .entry(facet_name)
-                .or_insert_with(Vec::new)
-                .push(FacetValue { value, count });
-        }
-
-        let facets: Vec<Facet> = facets_map
-            .into_iter()
-            .map(|(name, values)| Facet { name, values })
-            .collect();
-
-        Ok(facets)
     }
 
     /// Directly populates the content field since we use the ParadeDB BM25 index now
